keras/requirements.txt
numpy == 1.19.1
scikit-learn == 0.23.2
tensorflow-gpu == 1.15.0
opencv-python == 4.1.1
six == 1.15.0